Refactor EditLink to its own helper file
